requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
amadeus==9.0.0
apscheduler==3.10.4
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
plyer==2.1.0
jinja2==3.1.4
httpx==0.27.0
aiofiles==23.2.1
